add .float()

+++ predict_pv_yield/models/conv3d/model_sat_nwp.py
@@ -130,7 +130,7 @@ class Model(BaseModel):
 
         # ******************* Satellite imagery *************************
         # Shape: batch_size, channel, seq_length, height, width
-        sat_data = x.satellite.data
+        sat_data = x.satellite.data.float()
         batch_size, n_chans, seq_len, height, width = sat_data.shape
 
         # :) Pass data through the network :)
@@ -148,7 +148,7 @@ class Model(BaseModel):
 
         # add pv yield
         if self.include_pv_yield:
-            pv_yield_history = x[self.output_variable][:, : self.history_len_30 + 1].nan_to_num(nan=0.0)
+            pv_yield_history = x[self.output_variable][:, : self.history_len_30 + 1].nan_to_num(nan=0.0).float()
 
             pv_yield_history = pv_yield_history.reshape(
                 pv_yield_history.shape[0], pv_yield_history.shape[1] * pv_yield_history.shape[2]
@@ -160,7 +160,7 @@ class Model(BaseModel):
         if self.include_nwp:
 
             # shape: batch_size, n_chans, seq_len, height, width
-            nwp_data = x.nwp.data
+            nwp_data = x.nwp.data.float()
 
             out_nwp = F.relu(self.nwp_conv0(nwp_data))
             for i in range(0, self.number_of_conv3d_layers - 1):
